+++ lib/sj_sdk_testing.dart
@@ -5,20 +5,32 @@ import 'sj_sdk_testing_platform_interface.dart';
 enum Environment { dev, stg, uat, prod }
 
 class SjSdkTesting {
-  /// Provide your targeted environment: dev, stg, uat (default), prod
-  final Environment env;
+  SjSdkTesting._internal();
 
-  SjSdkTesting({
-    this.env = Environment.uat,
-  });
+  static final SjSdkTesting instance = SjSdkTesting._internal();
+
+  static Future<bool> init({
+    Environment env = Environment.uat,
+  }) async {
+    final isInitSuccess = await SjSdkTestingPlatform.instance.initEnvironment(env.toString());
+    return isInitSuccess;
+  }
 
   /// Sign in flow
-  Future<AuthModel> signIn() {
+  Future<AuthModel> signIn() async {
+    assert(
+      await SjSdkTestingPlatform.instance.isInitialized,
+      "Please call SjSdkTesting.init() before using this method",
+    );
     return SjSdkTestingPlatform.instance.signIn();
   }
 
   /// Sign up flow
-  Future<AuthModel> signUp() {
+  Future<AuthModel> signUp() async {
+    assert(
+      await SjSdkTestingPlatform.instance.isInitialized,
+      "Please call SjSdkTesting.init() before using this method",
+    );
     return SjSdkTestingPlatform.instance.signUp();
   }
 }